requirements.txt
dj-rest-auth==2.2.4
Django==4.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
djangorestframework-simplejwt==5.2.2
